+++ src/main/java/pl/plajer/buildbattle/arena/ArenaManager.java
@@ -40,6 +40,7 @@ import pl.plajer.buildbattle.arena.impl.SoloArena;
 import pl.plajer.buildbattle.handlers.ChatManager;
 import pl.plajer.buildbattle.handlers.PermissionManager;
 import pl.plajer.buildbattle.handlers.items.SpecialItem;
+import pl.plajer.buildbattle.handlers.party.GameParty;
 import pl.plajer.buildbattle.user.User;
 import pl.plajer.buildbattle.utils.Debugger;
 import pl.plajerlair.commonsbox.minecraft.misc.MiscUtils;
@@ -82,6 +83,29 @@ public class ArenaManager {
       player.sendMessage(plugin.getChatManager().getPrefix() + plugin.getChatManager().colorMessage("In-Game.Messages.Already-Playing"));
       return;
     }
+
+    //check if player is in party and send party members to the game
+    if (plugin.getPartyHandler().isPlayerInParty(player)) {
+      GameParty party = plugin.getPartyHandler().getParty(player);
+      if (party.getLeader().equals(player)) {
+        if (arena.getMaximumPlayers() - arena.getPlayers().size() >= party.getPlayers().size()) {
+          for (Player partyPlayer : party.getPlayers()) {
+            if (ArenaRegistry.getArena(partyPlayer) != null) {
+              if (ArenaRegistry.getArena(partyPlayer).getArenaState() == ArenaState.IN_GAME) {
+                continue;
+              }
+              leaveAttempt(partyPlayer, ArenaRegistry.getArena(partyPlayer));
+            }
+            partyPlayer.sendMessage(plugin.getChatManager().getPrefix() + plugin.getChatManager().formatMessage(arena,  plugin.getChatManager().colorMessage("In-Game.Messages.Join-As-Party-Member"), partyPlayer));
+            joinAttempt(partyPlayer, arena);
+          }
+        } else {
+          player.sendMessage(plugin.getChatManager().getPrefix() + plugin.getChatManager().formatMessage(arena,  plugin.getChatManager().colorMessage("In-Game.Messages.Lobby-Messages.Not-Enough-Space-For-Party"), player));
+          return;
+        }
+      }
+    }
+
     if (!plugin.getConfigPreferences().getOption(ConfigPreferences.Option.BUNGEE_ENABLED)) {
       if (!(player.hasPermission(PermissionManager.getJoinPerm().replace("<arena>", "*")) || player.hasPermission(PermissionManager.getJoinPerm().replace("<arena>", arena.getID())))) {
         player.sendMessage(plugin.getChatManager().getPrefix() + plugin.getChatManager().colorMessage("In-Game.Join-No-Permission")