+++ src/main/java/pl/plajer/buildbattle/arena/impl/SoloArena.java
@@ -421,7 +421,7 @@ public class SoloArena extends BaseArena {
     if (getVotingPlot() != null) {
       for (Player player : getPlayers()) {
         getVotingPlot().setPoints(getVotingPlot().getPoints() + getPlugin().getUserManager().getUser(player).getStat(StatsStorage.StatisticType.LOCAL_POINTS));
-        getPlugin().getUserManager().getUser(player).setStat(StatsStorage.StatisticType.LOCAL_POINTS, 0);
+        getPlugin().getUserManager().getUser(player).setStat(StatsStorage.StatisticType.LOCAL_POINTS, 3);
       }
     }
     voteRoutine();