+++ src/main/java/pl/plajer/buildbattle/utils/Utils.java
@@ -21,8 +21,8 @@ package pl.plajer.buildbattle.utils;
 import com.mojang.authlib.GameProfile;
 import com.mojang.authlib.properties.Property;
 
-import java.lang.reflect.Field;
 import java.lang.reflect.InvocationTargetException;
+import java.lang.reflect.Method;
 import java.util.LinkedHashMap;
 import java.util.LinkedList;
 import java.util.List;
@@ -98,13 +98,11 @@ public class Utils {
     SkullMeta headMeta = (SkullMeta) head.getItemMeta();
     GameProfile profile = new GameProfile(UUID.randomUUID(), null);
     profile.getProperties().put("textures", new Property("textures", url));
-    Field profileField;
     try {
-      profileField = headMeta.getClass().getDeclaredField("profile");
-      profileField.setAccessible(true);
-      profileField.set(headMeta, profile);
-
-    } catch (NoSuchFieldException | IllegalArgumentException | IllegalAccessException ignored) {
+      Method mtd = headMeta.getClass().getDeclaredMethod("setProfile", GameProfile.class);
+      mtd.setAccessible(true);
+      mtd.invoke(headMeta, profile);
+    } catch (NoSuchMethodException | InvocationTargetException | IllegalAccessException ignored) {
     }
 
     head.setItemMeta(headMeta);