+++ CHANGELOG.md
@@ -22,6 +22,7 @@ That's all, matcher will stop when detects next line started with `###` match
       * arenatype_pretty - prettified name of enum eg. Classic (instead SOLO)
  * Fixed VoteMenu NullPointerException
  * Added Parties support for PAF and Parties
+ * Now if user don´t vote the plot will get 3 points of this player to prevent unfair game behaviors
 
 ### 4.2.0 Release (16.11.2019 - 13.12.2019) (by Tigerpanzer)
 * Fixed arena deletion