+++ CHANGELOG.md
@@ -20,9 +20,10 @@ That's all, matcher will stop when detects next line started with `###` match
       * mapname - name of arena map
       * arenatype - raw enum name of arena (SOLO, TEAM etc)
       * arenatype_pretty - prettified name of enum eg. Classic (instead SOLO)
- * Fixed VoteMenu NullPointerException
- * Added Parties support for PAF and Parties
- * Now if user don´t vote the plot will get 3 points of this player to prevent unfair game behaviors
+* Fixed VoteMenu NullPointerException
+* Added Parties support for PAF and Parties
+* Now if user don´t vote the plot will get 3 points of this player to prevent unfair game behaviors
+* FIXED 1.15.X Head issues
 
 ### 4.2.0 Release (16.11.2019 - 13.12.2019) (by Tigerpanzer)
 * Fixed arena deletion